+++ main.go
@@ -1,30 +1,14 @@
 package main
 
 import (
-	"net/http"
-
 	"github.com/gin-gonic/gin"
+
+	"TeamVapp/AuthServer/auth"
 )
-
-type User struct {
-	Id       string `json:"id" binding:"required"`
-	Username string `json:"username" binding:"required"`
-	Password string `json:"password" binding:"required"`
-}
-
-func createUser(c *gin.Context) {
-	var newUser User
-	var err error = c.BindJSON(&newUser)
-	if err != nil {
-		c.JSON(http.StatusBadRequest, "Error parsing request body")
-		return
-	}
-	c.JSON(http.StatusCreated, newUser)
-}
 
 func main() {
 	router := gin.Default()
-	router.POST("/register", createUser)
+	router.POST("/register", auth.CreateUser)
 
 	router.Run("localhost:8080")
 }
